+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.4.1] - 2022-11-30
+
+### Fixed
+
+- Fix adding multiple user-functions
+
+
 ## [0.4.0] - 2022-11-30
 
 ### Added
@@ -18,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rule and RuleSet now return `Outcome` instances after evaluation that contain some metadata about the evaluation
 - `RuleSet::evaluate` does not halt rule evaluation when one rule returns `Err`
 - Bumped minimum supported Rust version to 1.62.1
+
 
 ## [0.3.0] - 2022-11-29
 
@@ -91,8 +100,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement a `RuleSet` as an entry point for executing a set of rules on a piece of data 
 - async `UserFunction` support
 
-[unreleased]: https://github.com/mendelt/reval/compare/v0.4.0...HEAD
-[0.3.0]: https://github.com/mendelt/reval/compare/v0.3.0...v0.4.0
+
+[unreleased]: https://github.com/mendelt/reval/compare/v0.4.1...HEAD
+[0.4.1]: https://github.com/mendelt/reval/compare/v0.4.0...v0.4.1
+[0.4.0]: https://github.com/mendelt/reval/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/mendelt/reval/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/mendelt/reval/compare/v0.1.3...v0.2.0
 [0.1.3]: https://github.com/mendelt/reval/compare/v0.1.2...v0.1.3